Tactics: exposing a way to read the error range bound, use it in `fail`s
When a tactic fails, it raises an exception, which does not read the
current error bound. The interpreter module running the tactic will then
bound the range by reading the current error bound, but it is too late
as it could have been popped away.

This moves the bounding logic into the `fail`s in the tactic module, to
read the bound immediately before raising the exception.

This is mostly useful for Pulse, as there is no way for normal userspace
clients to set the bound, so it wouldn't make a difference. Hence I'm
unsure about this patch.