Preserve timestamps when installing lib files
This way the timestamps of checked files end up more recent than the
sources, and in proper order between themselves too. This should not
matter, except that in externals runs of `fstar.exe --dep` we will get
rules for the files in the library too, usually wreaking havoc. This
avoids the problem... for now.