extraction: attempt to extract all type arguments
Using unit (MLTY_Erased) restricts the type arguments more than
needed in the cases where they actually are types, but this cannot be
determined at extraction time (e.g. they come from projecting a value
out of a record). In issue 3865, This causes the `f` project of `pak` to
receive a type of `pak -> unit myty`, which is unfolded to `pak -> unit
-> unit`, and all calls to this function become erased.

For Karamel extraction, retain the old behavior as an MLTY_Top is
outside of the Low* fragment. This is probably wrong and I would like to
revisit it. But Low* code should be safer from such bugs.

Fixes 3865.